stubs - reorganize - group services and models
