+++ stub/Services/BaseServiceStub.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: konstantin
+ * Date: 14.07.18
+ * Time: 0:30
+ */
+
+namespace MP\Services\Stub\Services;
+
+use MP\Services\BaseService;
+
+/**
+ * Class BaseServiceStub dummy service for testing
+ *
+ * @package MP\Services\Stub
+ */
+class BaseServiceStub extends BaseService
+{
+}
