+++ tests/CallTests.php
@@ -1,7 +1,7 @@
 <?php
 
-use MP\Services\Stub\BaseServiceStub;
-use MP\Services\Stub\ModelBaseServiceStub;
+use MP\Services\Stub\Services\BaseServiceStub;
+use MP\Services\Stub\Models\ModelBaseServiceStub;
 use PHPUnit\Framework\TestCase;
 use yii\base\UnknownMethodException;
