+++ tests/GetterTests.php
@@ -1,8 +1,8 @@
 <?php
 
-use MP\Services\Stub\BaseServiceStub;
-use MP\Services\Stub\ModelBaseServiceStub;
-use MP\Services\Stub\ModelNoServiceStub;
+use MP\Services\Stub\Services\BaseServiceStub;
+use MP\Services\Stub\Models\ModelBaseServiceStub;
+use MP\Services\Stub\Models\ModelNoServiceStub;
 use PHPUnit\Framework\TestCase;
 
 class GetterTests extends TestCase
